pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "full-self-crawl-agent"
version = "1.0.0"
description = "An AI-driven autonomous web crawling agent system"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "developer@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # LLM client dependencies
    "anthropic>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]
optional = [
    # Data processing
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    # HTML parsing
    "lxml>=4.9.0",
    # Image processing
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/username/full-self-crawl-agent"
Repository = "https://github.com/username/full-self-crawl-agent"

[project.scripts]
self-crawl-agent = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]